Test Bank/System Tests/AG2_Orchestration_Tests/requirements.txt
# Core testing framework
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-benchmark>=4.0.0
pytest-mock>=3.10.0
pytest-timeout>=2.1.0

# HTTP testing
requests>=2.28.0
aiohttp>=3.8.0
httpx>=0.24.0

# Mock and testing utilities
factory-boy>=3.2.0
faker>=18.0.0
freezegun>=1.2.0

# Performance testing
locust>=2.15.0
memory-profiler>=0.60.0
psutil>=5.9.0

# Security testing
bandit>=1.7.0
safety>=2.3.0
semgrep>=1.0.0

# Database testing
pytest-postgresql>=4.1.0
pytest-mock-resources>=2.6.0

# API testing
pytest-httpx>=0.22.0
responses>=0.23.0

# Documentation
pytest-html>=3.2.0
allure-pytest>=2.13.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
pre-commit>=3.0.0

# AG2 specific dependencies
sentence-transformers>=2.2.0
numpy>=1.21.0
PGvector>=0.4.0
openai>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
